// PortalNUR/src/models/enums/AreaCursoEnum.ts
export enum AreaCursoEnum {
    UNSET = 'Sin Seleccionar',
    ADMINISTRATIVAS = 'Ciencias Administrativas y Financieras',
    SOCIALES = 'Ciencias Sociales y Humanistas',
    JURIDICAS = 'Ciencias Jurídicas y Diplomáticas',
    TECNOLOGICAS = 'Ciencias Tecnológicas',
    SALUD = 'Ciencias de la Salud Humana',
}